Setup Boot key and enable tap dance
This isn't the double-tap miryoku implementation because in VIAL you
[cannot configure tap dance directly][1] as you would in QMK. Vial
[already defines `tap_dance_actions`][2] and it doesn't seem possible to
add to or modify that from a keymap easily.

Even though we can't set this up so that it implements miryoke exactly
out of the box, we can get close and then use the VIAL GUI to configure
tap dance the way we want with a double tap.

[1]: https://github.com/vial-kb/vial-qmk/issues/579
[2]: https://github.com/vial-kb/vial-qmk/blob/aac4e02152b131b050d7b6a8cab9a8f4b1ab989a/quantum/vial.c#L478