Add custom keycode for locking current layer
VIAL won't let us configure the tap dance directly in code, but TD(1) is
used with the expectation that it will be configured in VIAL to send the
LLOCK custom keycode on double tap.

The code for the layer lock feature was pulled from [this blog post][1]
by Pascal Getreuer.

---

The reason I'm adding a custom keycode instead of just using DF(layer)
is because with that approach the OLED doesn't correctly display the
active layer. With the custom LLOCK keycode implementation,
`get_highest_layer()` will allow us to show the correct layer name.

[1]: https://getreuer.info/posts/keyboards/layer-lock/index.html